Refactor: changed structure of js files execution.
