+++ index.html
@@ -46,7 +46,7 @@
                   <!-- Изменение размера изображения -->
                   <fieldset class="img-upload__scale  scale">
                     <button type="button" class="scale__control  scale__control--smaller">Уменьшить</button>
-                    <input type="text" class="scale__control  scale__control--value" value="55%" title="Image Scale" name="scale" readonly required>
+                    <input type="text" class="scale__control  scale__control--value" value="100%" title="Image Scale" name="scale" readonly required>
                     <button type="button" class="scale__control  scale__control--bigger">Увеличить</button>
                   </fieldset>
 
@@ -235,7 +235,5 @@
   </template>
   <script src="pristine/pristine.min.js"  type="text/javascript"></script>
   <script type="module" src="js/main.js"></script>
-  <script type="module" src="js/upload_form_animation.js"></script>
-  <script type="module" src="js/upload_form_validation.js"></script>
 </body>
 </html>
