+++ js/upload_form/upload_form_validation.js
@@ -9,22 +9,26 @@ const pristine = new Pristine(form, {
   errorTextTag: 'span',
   errorTextClass: 'form__error'
 },false);
-
 const hashtagRegexp = /^#[A-Za-zА-Яа-яЁё0-9]{1,19}$/;
 
-pristine.addValidator(commentInput, validateComment, 'Количество символов в комментарии должно быть от 20 до 140');
-function validateComment (value) {
-  return value.length > 20 && value.length < 140;
+export function activatePristineValidationOfUploadForm() {
+  pristine.addValidator(commentInput, validateComment, 'Количество символов в комментарии должно быть от 20 до 140');
+  pristine.addValidator(hashtagInput, validateHashtag, 'Хештег должен содержать от 1 до 19 символов и начинаться с решётки');
+  form.addEventListener('submit', (evt) => {
+    evt.preventDefault();
+    if(pristine.validate()){
+      form.submit();
+    }
+  });
+
 }
 
-pristine.addValidator(hashtagInput, validateHashtag, 'Хештег должен содержать от 1 до 19 символов и начинаться с решётки');
+function validateComment (value) {
+  return value.length > 20 && value.length < 140;
+
+}
+
 function validateHashtag (value) {
   return hashtagRegexp.test(value);
 }
 
-form.addEventListener('submit', (evt) => {
-  evt.preventDefault();
-  if(pristine.validate()){
-    form.submit();
-  }
-});
